abstract/modularize grant utils & minor logic enhancements
