+++ packages/abstraxion-core/src/index.ts
@@ -2,11 +2,12 @@ export { GranteeSignerClient } from "./GranteeSignerClient";
 export { AbstraxionAuth } from "./AbstraxionAuth";
 export { SignArbSecp256k1HdWallet } from "./SignArbSecp256k1HdWallet";
 export { StorageStrategy, RedirectStrategy } from "../src/types/strategyTypes";
-export { 
+export {
   GrantsResponse,
   Grant,
   TreasuryGrantConfig,
   SpendLimit,
   ContractGrantDescription,
-  DecodeAuthorizationResponse
+  DecodeAuthorizationResponse,
 } from "./types";
+export * from "@/utils/grant";
